requirements.txt
asgiref==3.8.1
Django==5.1.3
sqlparse==0.5.2
django-rest-framework==0.1.0
django-cors-headers==4.6.0
markdown==3.7
django-filter==24.3
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.26.4
